requirements.txt
anthropic>=0.40.0
pymupdf>=1.24.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
pytest>=8.0.0
